:recycle: Divide token integrations tests into multiple files
